microblog/requirements.txt
alembic==1.3.2
appdirs==1.4.2
appnope==0.1.0
Babel==2.3.4
backports-abc==0.5
beautifulsoup4==4.4.0
bleach==1.5.0
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
dominate==2.4.0
elasticsearch==7.1.0
Flask==1.1.1
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
html5lib==0.9999999
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
pyasn1==0.2.3
pycparser==2.17
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2016.10
requests==2.22.0
six==1.10.0
SQLAlchemy==1.3.12
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
